requirements.txt
numpy
pandas
pyarrow
PyYAML
sentence-transformers
matplotlib
seaborn
python-pptx
Pillow

